renamed polygon_utils2.py to polygon_utils.py
kept polygon_utils.py as polygon_utils_old.py
