+++ Users/Erin/all_e22_experiments.py
@@ -5,7 +5,7 @@ Also maps plot numbers to treatments.
 import os
 import re #regex
 
-from polygon_utils import *
+from polygon_utils_old import *
 import E22
 
 name = 'all_e22_experiments'
