+++ prog/find_plot.py
@@ -1,18 +1,22 @@
+import polygon_utils_old
 import polygon_utils
 
-
 def find_plot(df_row, rectangles):
-    # In the list rectangles, every element rectangles[i] is either
-    # a) a list of points representing corners of a polygon, or
-    # b) a function returning True iff df_row belongs to plot i.
+    # In the list or dict rectangles, every element rectangles[i] is either
+    # a) a list of points representing corners of a polygon,
+    # b) a Polygon object, or
+    # c) a function returning True iff df_row belongs to plot i.
     x = df_row.x
     y = df_row.y
     for i, p in rectangles.items():
         if callable(p):
             if p(df_row):
                 return i
+        elif isinstance(p, Polygon):
+            if p.contains(x,y):
+                return i
         else:
-            if polygon_utils.point_inside_polygon(x, y, p):
+            if polygon_utils_old.point_inside_polygon(x, y, p):
                 return i
     # so I am sure to recognize it even if I add one to make it
     # 1-based and then forget:
